insert 테스트 완료 * postgres insert 타입 에러 -> i64으로 변경 * redis 응답값에 newline이 섞어들어오는거 필더링

+++ collect/collect_server/src/worker/collect/client_list_worker.rs
@@ -5,7 +5,7 @@ use dbs_cmd;
 
 use parsing::redis_res::parsing_client_list;
 
-pub fn client_list_worker(link_key : u32, redis_conn : &'_ mut dbs::redis_pool::RedisRequester, pg_conn : &'_ mut dbs::pg_pool::PgUploader) -> Result<(),Box<dyn Error>> {
+pub fn client_list_worker(link_key : i32, redis_conn : &'_ mut dbs::redis_pool::RedisRequester, pg_conn : &'_ mut dbs::pg_pool::PgUploader) -> Result<(),Box<dyn Error>> {
     let cmd = dbs_cmd::REIDS_COMMANDLINE_MAP.get(&dbs_cmd::RedisCommand::ClientList).unwrap();
     let result = redis_conn.run_command(cmd, &[])?;
     
@@ -15,10 +15,10 @@ pub fn client_list_worker(link_key : u32, redis_conn : &'_ mut dbs::redis_pool::
     
     for client in list {
         pg_conn.execute(&pg_query, &[&link_key, 
-            &(client.id as u32), &client.addr, &(client.fd as u32), &client.name, &(client.age as u32),
-            &(client.idle as u32), &String::from(client.flags), &(client.db as u32), &(client.sub as u32),
-            &(client.psub as u32), &client.multi, &(client.qbuf as u32), &(client.qbuf_free as u32),
-            &(client.obl as u32), &(client.oll as u32), &(client.omem as u32), &String::from(client.events), &client.cmd])?;
+            &client.id, &client.addr, &client.fd, &client.name, &client.age,
+            &client.idle, &String::from(client.flags), &client.db, &client.sub,
+            &client.psub, &(client.multi as i64), &client.qbuf, &client.qbuf_free,
+            &client.obl, &client.oll, &client.omem, &String::from(client.events), &client.cmd])?;
     }
 
     Ok(())
